pyproject.toml
[tool.poetry]
name = "newtimetable"
version = "0.1.0"
description = ""
authors = ["nagraver <ttolla25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.4.1"
asyncio = "^3.4.3"
pymongo = "^4.6.2"
bs4 = "^0.0.2"
requests = "^2.31.0"
lxml = "^5.1.0"
sentry-sdk = "^1.44.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
